Issue #846 - save filter on check checkbox

+++ app/app/scripts/services/Filter.js
@@ -54,8 +54,8 @@ angular.module('hearth.services').factory('Filter', [
             apply: function(filterData, save, applySave){
 
                 $location.search(filterData);
-                if(applySave) {
-                    if (save && $rootScope.loggedUser._id) {
+                if(applySave && $rootScope.loggedUser._id) {
+                    if (save) {
                         this.setUserFilter(filterData);
                     } else {
                         this.deleteUserFilter();
